requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Blockchain and Web3
web3==6.15.1
eth-account==0.10.0

# Firebase (compatible version)
firebase-admin==6.4.0

# HTTP requests
requests==2.31.0

# Logging and utilities
colorama==0.4.6
